move resources file to OpenNumismat to fix packaging (deb, arch)
