+++ OpenNumismat/main.py
@@ -10,8 +10,8 @@ from PyQt5.QtCore import QCoreApplication, QTranslator, QUrl, QUrlQuery, PYQT_VE
 from PyQt5.QtWidgets import QApplication, QMessageBox
 from PyQt5.QtGui import QDesktopServices
 
-import resources
 import OpenNumismat
+import OpenNumismat.resources
 from OpenNumismat.Settings import Settings
 from OpenNumismat.MainWindow import MainWindow
 from OpenNumismat.Tools import TemporaryDir
