+++ OpenNumismat/resources.py
